dtype issue with torchinfo

+++ clinicadl/models/vanilla.py
@@ -200,7 +200,10 @@ class VanillaModel(Model):
         from torchinfo import summary
 
         summary_ = summary(
-            self.network, input_data=input_data.get_field(IMAGE), batch_dim=0, verbose=0
+            self.network,
+            input_data=input_data.get_field(IMAGE, dtype=torch.float32),
+            batch_dim=0,
+            verbose=0,
         )
 
         return str(summary_)
